requirements-dev.txt
# Development dependencies
boto3>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
python-dotenv>=1.0.0
